This is about **data** format.
Let's not get ahead of ourself, this format is for **data** only.
Anything else we might send to the client can be discovered through the API config.

Yes, that means we will have a different format for answering the client.
It will be generated from this one, by giving parameters from the API to the XSLT for the missing
information (controls/links, meta, error).